+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
   * added possibility to force a method to throw an exception using `bovigo\callmap\throws()`
   * added `bovigo\callmap\NewInstance::classname()` to retrieve the name of the generated proxy
+  * added `bovigo\callmap\onConsecutiveCalls()` for passing a list of invocation results, removing api status from `bovigo\callmap\InvocationResults`
 
 
 0.1.0 (2015-04-09)